+++ packages/rocket-fileuploader-aws-infrastructure/package.json
@@ -1,7 +1,7 @@
 {
-  "name": "@boostercloud/rocket-uploads3-store-event-aws-infrastructure",
+  "name": "@boostercloud/rocket-fileuploader-aws-infrastructure",
   "version": "0.9.5699",
-  "description": "Booster rocket to upload file in S3 and store event in Event Store when it happens.",
+  "description": "Booster rocket to upload files in File Storage. A new event will be stored in the Events Store as a consequence.",
   "keywords": [
     "Booster",
     "plugin",
@@ -26,7 +26,7 @@
     "@aws-cdk/aws-dynamodb": "1.67.0",
     "@aws-cdk/aws-lambda": "1.67.0",
     "@aws-cdk/aws-lambda-event-sources": "1.67.0",
-    "@boostercloud/framework-types": "^0.10.0",
+    "@boostercloud/framework-types": "^0.11.1",
     "@aws-cdk/core": "1.67.0",
     "aws-sdk": "2.764.0"
   },
@@ -44,6 +44,13 @@
     "url": "https://github.com/boostercloud/booster/issues"
   },
   "devDependencies": {
-    "@boostercloud/framework-provider-aws-infrastructure": "^0.10.0"
+    "@boostercloud/framework-provider-aws-infrastructure": "^0.11.1",
+    "@types/faker": "5.1.5",
+    "chai": "4.2.0",
+    "chai-as-promised": "7.1.1",
+    "faker": "5.1.0",
+    "sinon": "9.2.3",
+    "sinon-chai": "3.5.0",
+    "velocityjs": "^2.0.0"
   }
 }
