+++ packages/rocket-fileuploader-aws-infrastructure/src/index.ts
