+++ packages/rocket-fileuploader-aws-infrastructure/src/lambdas/package.json
