+++ packages/rocket-fileuploader-aws-infrastructure/src/s3-notification-event-store.ts
@@ -5,6 +5,7 @@ import { Function, Runtime, Code } from '@aws-cdk/aws-lambda'
 import { Table } from '@aws-cdk/aws-dynamodb'
 import { S3EventSource } from '@aws-cdk/aws-lambda-event-sources'
 import { createPolicyStatement } from '@boostercloud/framework-provider-aws-infrastructure/dist/infrastructure/stacks/policies'
+import { BoosterConfig } from '@boostercloud/framework-types'
 
 const path = require('path')
 
@@ -15,10 +16,8 @@ export type AWSS3EventsLambdaParams = {
 }
 
 export class S3NotificationEventStoreStack {
-  public static mountStack(params: AWSS3EventsLambdaParams, stack: Stack): void {
-    const appName = stack.stackName.split('-app')[0]
+  public static mountStack(params: AWSS3EventsLambdaParams, stack: Stack, config: BoosterConfig): void {
 
-    // Create new bucket to upload file
     const sourceBucket = new Bucket(stack, 'sourceUploadBucket' + params.entityTypeName, {
       bucketName: params.bucketName,
       removalPolicy: RemovalPolicy.DESTROY,
@@ -26,35 +25,28 @@ export class S3NotificationEventStoreStack {
 
     const eventsStore = stack.node.tryFindChild('events-store') as Table
 
-    // Crete new Lambda function to be triggered once new file is uploaded in bucket
-    const processorFunction = new Function(stack, 'rocketS3Trigger' + params.entityTypeName, {
+    const fileTriggerFunction = new Function(stack, 'rocketS3Trigger' + params.entityTypeName, {
       runtime: Runtime.NODEJS_12_X,
       timeout: Duration.minutes(15),
       memorySize: 1024,
       handler: 'index.handler',
-      functionName: appName + '-' + params.entityTypeName.toLowerCase() +'-s3-rocket-trigger',
+      functionName: config.appName + '-' + params.entityTypeName.toLowerCase() + '-s3-rocket-trigger',
       code: Code.fromAsset(path.join(__dirname, 'lambdas')),
       environment: {
-        EVENT_STORE_NAME:eventsStore.tableName,
+        EVENT_STORE_NAME: eventsStore.tableName,
         ENTITY_TYPE_NAME: params.entityTypeName,
         TYPE_NAME: params.eventTypeName,
       },
     })
 
-    // Grant access for this lambda to persist new records in the Events Store
-    processorFunction.addToRolePolicy(
-      createPolicyStatement([eventsStore.tableArn], ['dynamodb:Put*'])
-    )
-    // Grant access for this lambda to read from the S3 bucket
-    sourceBucket.grantRead(processorFunction)
+    fileTriggerFunction.addToRolePolicy(createPolicyStatement([eventsStore.tableArn], ['dynamodb:Put*']))
+    sourceBucket.grantRead(fileTriggerFunction)
 
-    // Create trigger from S3 to Lambda
     const uploadEvent = new S3EventSource(sourceBucket, {
       events: [EventType.OBJECT_CREATED],
     })
-    processorFunction.addEventSource(uploadEvent)
+    fileTriggerFunction.addEventSource(uploadEvent)
 
-    // Grant the Events handler lambda access to read from S3 bucket
     const eventsHandlerLambda = stack.node.tryFindChild('events-main') as Function
     eventsHandlerLambda.addToRolePolicy(
       createPolicyStatement([sourceBucket.bucketArn, sourceBucket.bucketArn + '/*'], ['s3:*'])
