+++ packages/rocket-fileuploader-aws-infrastructure/tsconfig.eslint.json
