+++ packages/rocket-fileuploader-aws-infrastructure/tsconfig.json
